Revert "Merge pull request #455 from vtex-apps/fix/mobile-filter"
This reverts commit 98dec5ddc5814626d0e93c9d86519e43e93e0709, reversing
changes made to 66830edfa76014e0c32d23a5b1af7559f691d535.
